requirements.txt
asgiref==3.8.1
Django==4.2.13
django-filter==24.2
django-jazzmin==3.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
PyJWT==2.8.0
sqlparse==0.5.0
typing-extensions==4.12.2
tzdata==2024.1
